categorize tests and improve coverage
